fix(RREL-027): split relayServerUtils functions into multiple files
