+++ src/relayServerUtils/constants.ts
@@ -1,0 +1,12 @@
+export const TRANSFER_HASH = 'a9059cbb';
+export const TRANSFER_FROM_HASH = '23b872dd';
+const INDEX_WHERE_VALUE_STARTS_ON_TRANSFER = 74;
+const INDEX_WHERE_VALUE_STARTS_ON_TRANSFER_FROM = 138;
+export const US_DOLLAR_SYMBOL = 'USD';
+
+export const VALUE_INDEX_IN_DATA = new Map<string, number>();
+VALUE_INDEX_IN_DATA.set(TRANSFER_HASH, INDEX_WHERE_VALUE_STARTS_ON_TRANSFER);
+VALUE_INDEX_IN_DATA.set(
+  TRANSFER_FROM_HASH,
+  INDEX_WHERE_VALUE_STARTS_ON_TRANSFER_FROM
+);
