+++ src/relayServerUtils/getMethodHashFromData.ts
@@ -1,0 +1,3 @@
+export function getMethodHashFromData(data: string) {
+  return data.substring(2, 10);
+}
